+++ src/utils/browser.ts
@@ -33,3 +33,7 @@ export const clearUserSessionKeys = () => {
     clearStorageByMatchTerm(item);
   }
 };
+
+export const isTouchDevice = () => {
+  return 'ontouchstart' in window || navigator.maxTouchPoints > 0 || navigator.maxTouchPoints > 0;
+};